refactor(repair): share comment compaction
